+++ src/main/java/com/super_bits/modulosSB/Persistencia/dao/calculosListagens/ItfListasJPA.java
@@ -65,7 +65,7 @@ public interface ItfListasJPA extends ItfListas {
             if (criterioQuery == null) {
                 throw new UnsupportedOperationException("Impossivel determinar o tipo de pesquisa padrão a ser realizada");
             }
-            Map<String, Integer> valoresParametro = new HashMap<>();
+            Map<String, Long> valoresParametro = new HashMap<>();
             List<Predicate> condicoes = new ArrayList<>();
             int i = 0;
             for (Object parametro : pParametros) {
